refactor: move _paginate from tools/_helpers to utils
tools/_helpers imports zotero_mcp.client, so client.py and
semantic_search.py could not use _paginate without a circular import.
utils.py has no internal imports, making it a safe home. _helpers
re-exports the name so existing _helpers._paginate references keep
working.
